refactor organization for Projects components
